Allow database ids as valid slug for tags.
Bevor this change it was for example not possible to use a product id
as name (slug) for a tag, because the application interpreted the id as
a tag id, instead of a tag name.

A valid use case would be to display a link to a product's related products.
Allowing a product's id as tag name (slug) it is easy to build an URL to related products e.g.
product/tag/<product_id> .

The old behaviour is still in service as a fallback.